read from taxons file
